Inconsistent merge state: could not find file MERGE_MSG in .git folder.
This file contains the commit message for a merge commit.

+++ src/main/java/com/project/market/customerService/controller/CustomerServiceController.java
@@ -12,12 +12,15 @@ import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.project.market.customerService.model.vo.FrequentlyQuestion;
 import com.project.market.customerService.model.vo.Proposal;
 import com.project.market.customerService.model.vo.Question;
 import com.project.market.product.model.vo.Product;
 import com.project.market.security.model.vo.Member;
 import org.apache.ibatis.session.RowBounds;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -35,6 +38,7 @@ import lombok.extern.slf4j.Slf4j;
 @Controller
 @Slf4j
 @RequestMapping("/service")
+@PropertySource("classpath:filePath.properties")
 public class CustomerServiceController {
 
     @Autowired
@@ -43,11 +47,30 @@ public class CustomerServiceController {
     @Autowired
     private ServletContext application;
 
+    @Value("${board.anno.path}")
+    private String annoPath;
+
+    @Value("${board.ques.path}")
+    private String quesPath;
+
+    @Value("${board.prpr.path}")
+    private String prprPath;
+
+    @Value("${board.ecpr.path}")
+    private String ecprPath;
+
+    @Value("${board.lapr.path}")
+    private String laprPath;
+
+    @Value("${board.frqu.path}")
+    private String frquPath;
+
     private final String QUES = "question";
     private final String ANNO = "announcement";
     private final String PRPR = "productProposal";
     private final String ECPR = "echoProposal";
     private final String LAPR = "largeProposal";
+    private final String FRQU = "frequentlyQuestion";
     private final String PRRE = "productReview";
 
     @GetMapping("/view/{boardId}")
@@ -107,6 +130,16 @@ public class CustomerServiceController {
 
                     model.addAttribute("largeProposal", laprList);
                     break;
+                case FRQU:
+                    totalContent = customerServiceService.countAllFrequentlyQuestion();
+                    log.debug("totalContent = {}", totalContent);
+
+                    commonThings = commonUtils(cPage, totalContent, request);
+                    List<FrequentlyQuestion> frquList = customerServiceService.selectAllFrequentlyQuestion((RowBounds) commonThings.get("rowBounds"));
+                    log.debug("frequently question list = {}", frquList);
+
+                    model.addAttribute("frequentlyQuestion", frquList);
+                    break;
             }
             model.addAttribute("pagebar", (String) commonThings.get("pagebar"));
         }catch(Exception e){
@@ -133,6 +166,12 @@ public class CustomerServiceController {
                         model.addAttribute("announce", announce);
                     }
                     break;
+                case FRQU:
+                    if("modify".equals((String) param.get("type"))){
+                        FrequentlyQuestion frequently = customerServiceService.selectOneFrequentlyQuestion(param);
+                        model.addAttribute("frequentlyQuestion", frequently);
+                    }
+                    break;
             }
 
         }catch(Exception e){
@@ -143,9 +182,8 @@ public class CustomerServiceController {
     }
 
     @PostMapping("/modify/{boardId}")
-    public String modifyBoard(@PathVariable(required = true) String boardId, Question question, Proposal proposal, Announcement announcement,
+    public String modifyBoard(@PathVariable(required = true) String boardId, Question question, Proposal proposal, Announcement announcement, FrequentlyQuestion frequence,
                               @RequestParam(name = "upFile", required = false) MultipartFile[] upFiles, RedirectAttributes redirectAttr) throws IOException {
-        String path = "";
         String directory = "";
         Map<String, Object> boardCode = new HashMap<>();
         List<Attachment> newAttachments = new ArrayList<>();
@@ -154,13 +192,12 @@ public class CustomerServiceController {
             switch (boardId){
                 case QUES:
                     log.debug("question = {}", question);
-                    path = "/resources/upload/question/";
-                    directory = application.getRealPath(path);
+                    directory = application.getRealPath(quesPath);
                     boardCode.put("code", question.getQCode());
                     Question oldQuestion = customerServiceService.selectOneQuestion(boardCode);
-                    if(upFiles.length > 0){
+                    if(upFiles != null && upFiles.length > 0){
                         if(!oldQuestion.getAttachments().isEmpty()){
-                            deleteAttachment(oldQuestion.getAttachments(), path);
+                            deleteAttachment(oldQuestion.getAttachments(), quesPath);
                             customerServiceService.deleteAttachments(boardCode);
                         }
 
@@ -176,13 +213,12 @@ public class CustomerServiceController {
                     break;
                 case ANNO:
                     log.debug("announcement = {}", announcement);
-                    path = "/resources/upload/announce/";
-                    directory = application.getRealPath(path);
+                    directory = application.getRealPath(annoPath);
                     boardCode.put("code", announcement.getAnCode());
                     Announcement oldAnnounce = customerServiceService.selectOneAnnouncement(boardCode);
-                    if(upFiles.length > 0){
+                    if(upFiles != null && upFiles.length > 0){
                         if(!oldAnnounce.getAttachments().isEmpty()){
-                            deleteAttachment(oldAnnounce.getAttachments(), path);
+                            deleteAttachment(oldAnnounce.getAttachments(), annoPath);
                             customerServiceService.deleteAttachments(boardCode);
                         }
 
@@ -195,6 +231,27 @@ public class CustomerServiceController {
                     customerServiceService.modifyAnnouncement(announcement);
 
                     redirectAttr.addFlashAttribute("msg", "공지 수정 완료!");
+                    break;
+                case FRQU:
+                    log.debug("frequence = {}", frequence);
+                    directory = application.getRealPath(frquPath);
+                    boardCode.put("code", frequence.getCode());
+                    FrequentlyQuestion oldFrequence = customerServiceService.selectOneFrequentlyQuestion(boardCode);
+                    if(upFiles != null && upFiles.length > 0){
+                        if(!oldFrequence.getAttachments().isEmpty()){
+                            deleteAttachment(oldFrequence.getAttachments(), frquPath);
+                            customerServiceService.deleteAttachments(boardCode);
+                        }
+
+                        newAttachments = commonAttachment(upFiles, directory);
+                    }
+
+                    if(!newAttachments.isEmpty())
+                        frequence.setAttachments(newAttachments);
+
+                    customerServiceService.modifyFrequentlyQuestion(frequence);
+
+                    redirectAttr.addFlashAttribute("msg", "자주하는 질문 수정 완료!");
                     break;
             }
         }catch(Exception e){
@@ -250,6 +307,11 @@ public class CustomerServiceController {
                     log.debug("largeProposal", proposal);
                     model.addAttribute("largeProposal", proposal);
                     break;
+                case FRQU:
+                    FrequentlyQuestion frequence = customerServiceService.selectOneFrequentlyQuestion(boardCode);
+                    log.debug("frequence", frequence);
+                    model.addAttribute("frequentlyQuestion", frequence);
+                    break;
             }
         }catch(Exception e){
             log.error(e.getMessage(), e);
@@ -261,51 +323,52 @@ public class CustomerServiceController {
     @PostMapping("/delete/{boardId}")
     public String deleteBoard(@PathVariable(required = true) String boardId, @RequestParam String code, RedirectAttributes redirectAttr){
         try{
-            String path = "";
             Map<String, Object> boardCode = new HashMap<>();
             boardCode.put("code", code);
             List<Attachment> attachments = customerServiceService.selectAllAttachments(boardCode);
 
             switch (boardId){
                 case ANNO:
-                    path = "/resources/upload/announce/";
-                    deleteAttachment(attachments, path);
+                    deleteAttachment(attachments, annoPath);
 
                     customerServiceService.deleteAnnouncement(boardCode);
 
                     redirectAttr.addFlashAttribute("msg", "공지사항 삭제 성공");
                     break;
                 case QUES:
-                    path = "/resources/upload/question/";
-                    deleteAttachment(attachments, path);
+                    deleteAttachment(attachments, quesPath);
 
                     customerServiceService.deleteMyQuestion(boardCode);
 
                     redirectAttr.addFlashAttribute("msg", "질문 삭제 성공");
                     break;
                 case PRPR:
-                    path = "/resources/upload/productProposal/";
-                    deleteAttachment(attachments, path);
+                    deleteAttachment(attachments, prprPath);
 
                     customerServiceService.deleteOneProductProposal(boardCode);
 
                     redirectAttr.addFlashAttribute("msg", "제안 삭제 성공");
                     break;
                 case ECPR:
-                    path = "resources/upload/echoProposal/";
-                    deleteAttachment(attachments, path);
+                    deleteAttachment(attachments, ecprPath);
 
                     customerServiceService.deleteOneEchoProposal(boardCode);
 
                     redirectAttr.addFlashAttribute("msg", "에코포장 피드백 삭제 성공");
                     break;
                 case LAPR:
-                    path = "resources/upload/largeProposal/";
-                    deleteAttachment(attachments, path);
+                    deleteAttachment(attachments, laprPath);
 
                     customerServiceService.deleteOneLargeProposal(boardCode);
 
                     redirectAttr.addFlashAttribute("msg", "대량구매 문의 삭제 성공");
+                    break;
+                case FRQU:
+                    deleteAttachment(attachments, frquPath);
+
+                    customerServiceService.deleteOneFrequentlyQuestion(boardCode);
+
+                    redirectAttr.addFlashAttribute("msg", "자주하는 문의 삭제 성공");
                     break;
             }
         }catch(Exception e){
@@ -316,11 +379,12 @@ public class CustomerServiceController {
     }
 
     @PostMapping("/enroll/{boardId}")
-    public String enrollBoard(@PathVariable(required = true) String boardId, Proposal proposal, Announcement announcement, Question question, @RequestParam(name="upFile", required = false) MultipartFile[] upFiles,
+    public String enrollBoard(@PathVariable(required = true) String boardId, Proposal proposal, FrequentlyQuestion frequence, Announcement announcement, Question question, @RequestParam(name="upFile", required = false) MultipartFile[] upFiles,
                               RedirectAttributes redirectAttr) throws IOException {
         log.debug("proposal = {}", proposal);
         log.debug("announcement = {}", announcement);
         log.debug("question = {}", question);
+        log.debug("upfiles = {}", upFiles);
 
         List<Attachment> attachments = new ArrayList<>();
         String saveDirectory = "";
@@ -328,9 +392,9 @@ public class CustomerServiceController {
             switch (boardId){
                 case ANNO:
                     log.debug("announcement = {}", announcement);
-                    saveDirectory = application.getRealPath("/resources/upload/announce");
+                    saveDirectory = application.getRealPath(annoPath);
 
-                    if(upFiles.length > 0)
+                    if(upFiles != null && upFiles.length > 0)
                         attachments = commonAttachment(upFiles, saveDirectory);
                     if(!attachments.isEmpty())
                         announcement.setAttachments(attachments);
@@ -339,9 +403,9 @@ public class CustomerServiceController {
                     break;
                 case QUES:
                     log.debug("question = {}", question);
-                    saveDirectory = application.getRealPath("/resources/upload/question");
+                    saveDirectory = application.getRealPath(quesPath);
 
-                    if(upFiles.length > 0)
+                    if(upFiles != null && upFiles.length > 0)
                         attachments = commonAttachment(upFiles, saveDirectory);
                     if(!attachments.isEmpty())
                         question.setAttachments(attachments);
@@ -350,9 +414,9 @@ public class CustomerServiceController {
                     break;
                 case PRPR:
                     log.debug("productProposal = {}", proposal);
-                    saveDirectory = application.getRealPath("/resources/upload/productProposal");
+                    saveDirectory = application.getRealPath(prprPath);
 
-                    if(upFiles.length > 0)
+                    if(upFiles != null && upFiles.length > 0)
                         attachments = commonAttachment(upFiles, saveDirectory);
                     if(!attachments.isEmpty())
                         proposal.setAttachments(attachments);
@@ -361,9 +425,9 @@ public class CustomerServiceController {
                     break;
                 case ECPR:
                     log.debug("echoProposal = {}", proposal);
-                    saveDirectory = application.getRealPath("/resources/upload/echoProposal");
+                    saveDirectory = application.getRealPath(ecprPath);
 
-                    if(upFiles.length > 0)
+                    if(upFiles != null && upFiles.length > 0)
                         attachments = commonAttachment(upFiles, saveDirectory);
                     if(!attachments.isEmpty())
                         proposal.setAttachments(attachments);
@@ -372,14 +436,25 @@ public class CustomerServiceController {
                     break;
                 case LAPR:
                     log.debug("largeProposal = {}", proposal);
-                    saveDirectory = application.getRealPath("/resources/upload/largeProposal");
+                    saveDirectory = application.getRealPath(laprPath);
 
-                    if(upFiles.length > 0)
+                    if(upFiles != null && upFiles.length > 0)
                         attachments = commonAttachment(upFiles, saveDirectory);
                     if(!attachments.isEmpty())
                         proposal.setAttachments(attachments);
                     log.debug("largeProposal + attachments = {}", proposal);
                     customerServiceService.insertLargeProposal(proposal);
+                    break;
+                case FRQU:
+                    log.debug("frequence = {}", frequence);
+                    saveDirectory = application.getRealPath(frquPath);
+
+                    if(upFiles != null && upFiles.length > 0)
+                        attachments = commonAttachment(upFiles, saveDirectory);
+                    if(!attachments.isEmpty())
+                        frequence.setAttachments(attachments);
+                    log.debug("frequently question + attachments = {}", frequence);
+                    customerServiceService.insertFrequentlyQuestion(frequence);
                     break;
             }
             redirectAttr.addFlashAttribute("msg", "게시물 등록 성공!");
